split responsabilities
