+++ connectors/src/connectors/slack/lib/messages.ts
@@ -2,12 +2,12 @@ import {
   getBotOrUserName,
   getUserInfo,
 } from "@connectors/connectors/slack/lib/bot_user_helpers";
-import { getChannelNameById } from "@connectors/connectors/slack/lib/channels";
 import { splitSlackAttachments } from "@connectors/connectors/slack/lib/message_attachments";
 import {
   EMPTY_SECTION,
   formatSlackMessageForLLM,
 } from "@connectors/connectors/slack/lib/message_formatter";
+import { processMessageForMentions } from "@connectors/connectors/slack/lib/message_mentions";
 import type { CoreAPIDataSourceDocumentSection } from "@connectors/lib/data_sources";
 import { renderDocumentTitleAndContent } from "@connectors/lib/data_sources";
 import { formatDateForUpsert } from "@connectors/lib/formatting";
@@ -15,48 +15,6 @@ import type { DataSourceConfig, ModelId } from "@connectors/types";
 import { safeSubstring } from "@connectors/types";
 import type { WebClient } from "@slack/web-api";
 import type { MessageElement } from "@slack/web-api/dist/types/response/ConversationsRepliesResponse";
-
-// `<@U123>` and `<#C123>`: Slack sends mentions as bare ids (no label) in both rich_text and
-// mrkdwn, and the formatter leaves these tokens intact for us to resolve here.
-const USER_MENTION_RE = /<@[UW][A-Z0-9]+>/g;
-const CHANNEL_MENTION_RE = /<#[A-Z0-9]+>/g;
-
-// Resolves a single mention token to its display form via the Slack API: `<@U123>` -> `@name`,
-// `<#C123>` -> `#name`, falling back to the raw id when it does not resolve.
-async function resolveMentionToken(
-  token: string,
-  connectorId: ModelId,
-  slackClient: WebClient
-): Promise<string> {
-  const id = token.replace(/[<@#>]/g, "");
-  if (token.startsWith("<@")) {
-    const { name } = await getUserInfo(id, connectorId, slackClient);
-    return `@${name ?? id}`;
-  }
-  const name = await getChannelNameById(connectorId, slackClient, id);
-  return `#${name ?? id}`;
-}
-
-// Resolves every user/channel mention token the formatter left in the rendered body.
-async function processMessageForMentions(
-  body: string,
-  connectorId: ModelId,
-  slackClient: WebClient
-): Promise<string> {
-  const tokens = new Set([
-    ...(body.match(USER_MENTION_RE) ?? []),
-    ...(body.match(CHANNEL_MENTION_RE) ?? []),
-  ]);
-
-  let resolved = body;
-  for (const token of tokens) {
-    resolved = resolved.replaceAll(
-      token,
-      await resolveMentionToken(token, connectorId, slackClient)
-    );
-  }
-  return resolved;
-}
 
 // Single entry point to turn a Slack message into its upsert body text: renders the
 // content (blocks first, top-level `text` as a fallback), appends forwarded (unfurl)
